js/xpconnect/tests/chrome/test_bug1430164.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1430164
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1430164</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript">
  /** Test for Bug 1430164 **/
      var Cu = Components.utils;
      var s = new Cu.Sandbox(window, { sandboxPrototype: window } );
      var t;
      var desc;
      try {
          t = Cu.evalInSandbox('Node === Node', s);
          ok(t, "Object identity should be sane");

          t = Cu.evalInSandbox('Node === this.Node', s);
          ok(t, "Node should match this.Node in toplevel");

          t = Cu.evalInSandbox('Node === window.Node', s);
          ok(t, "Node should match window.Node");
      } catch (e) {
          ok(false, "Should not get an exception: " + e);
      }
  </script>
</head>
</html>
